+++ src/sync/mod.rs
@@ -1,18 +1,3 @@
-use std::fmt::Display;
-
+//! A set of safe synchronization primitives for high-performance multi-threaded systems
+//! with various performance vs. accuracy trade offs.
 pub mod gvt;
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum ComputeLayout {
-    HubSpoke,
-    Decentralized,
-}
-
-impl Display for ComputeLayout {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ComputeLayout::HubSpoke => write!(f, "Centralized/Master Channel"),
-            ComputeLayout::Decentralized => write!(f, "Decentralized"),
-        }
-    }
-}
